// src/app/interfaces/search-information.ts
export interface SearchInformation {
  language: string;
  searchQuery: string;
  numOfRecords: number;
  mediaType: string,
  selectedChannel: string;
  selectedNewspaper: string;
  selectedRadio: string;
  selectedMagazine: string;
  selectedSocialMedia: string;
  country: string;
  stateProvince: string;
  city: string;
  startDateTime: string;
  endDateTime: string;
  startDate: string;
  endDate: string;
  startTime: string;
  endTime: string;
  realTimeAlerts: string;
  emailAddresses: string;
  reportFormats: string;
}